% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abun_assign.R
\name{abun_assign}
\alias{abun_assign}
\title{Breeding assignment using weighted stable isotope and abundance data}
\usage{
abun_assign(iso_data, rel_abun, iso_weight, abun_weight, odds = 0.67)
}
\arguments{
\item{iso_data}{Dataframe containing the isotope-based assignment results from the iso_assign() function}

\item{rel_abun}{Vector of relative abundance values for each breeding cell (should sum to 1).}

\item{iso_weight}{Weight value to apply to stable isotope data.}

\item{abun_weight}{Weight value to apply to relative abundance data.}

\item{odds}{Odds for determining likely/unlikely origins. Default is 0.67.}
}
\value{
Data frames containing the isotope-based assignment results plus the weighted posterior probabilities & origins for each individual
}
\description{
Function to estimate the posterior probability of origin and likely/unlikely origins for stable hydrogen isotope samples and relative breeding abundance
}
\examples{
woth_iso <- iso_assign(dd = woth_dd, df_base = woth_base$ddf, lat = woth_base$y, lon = woth_base$x)
abun_assign(iso_Data = woth_iso, rel_abun = woth_base$rel.abun, iso_weight = -0.7, abun_weight = 0)
}
\keyword{abundance}
\keyword{assignment;}
\keyword{isotopes;}
\keyword{stable}

