% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/area_assign.R
\name{area_assign}
\alias{area_assign}
\title{Estimate the assignment area for known-origin tissue samples}
\usage{
area_assign(summ, iso = TRUE)
}
\arguments{
\item{summ}{Assignment results from either the iso_assign() or abun_assign() functions}

\item{iso}{Should area be estimated for the isotope-only assignment (default) or with abundance used a prior}
}
\value{
Assignment area (i.e., the mean number of cells identified as likely for all individuals)
}
\description{
Function to estimate the mean number of cells identified as likely for all individuals in a given sample
}
\examples{
area_assign(summ = woth_iso)
}
\keyword{area}
\keyword{assignment}
\keyword{assignment;}
\keyword{isotopes;}
\keyword{stable}

