% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_rate.R
\name{error_rate}
\alias{error_rate}
\title{Estimate the assignment error rate for known-origin tissue samples}
\usage{
error_rate(origin, origin.cell)
}
\arguments{
\item{origin}{Matrix with columns indicating likely/unlikely cells for each individual.}

\item{origin.cell}{Vector indicating the true breeding cell (i.e. row in origin matrix) for each individual.}
}
\value{
Error rate (i.e., the proportion of individuals that are miss classified)
}
\description{
Function to estimate the proportion of individuals that are incorrectly assigned to their actual breeding origin
}
\examples{
error_rate(origin = amre_assign$iso.origin, origin.cell = amre$origin)
}
\keyword{assignment;}
\keyword{error}
\keyword{isotopes;}
\keyword{rate}
\keyword{stable}

