% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iso_assign.R
\name{iso_assign}
\alias{iso_assign}
\title{Breeding assignment using stable isotope data}
\usage{
iso_assign(dd, df.base, odds = 3)
}
\arguments{
\item{dd}{Vector of length = # individuals containing the feather deuterium values.}

\item{df.base}{Vector of predicted deuterium values for each potential breeding location.}

\item{odds}{Denominator of the odds ratio for determining likely/unlikely origins. Default is 3.}
}
\value{
A list containing three dataframes:
 iso.like contains a likelihood values for each breeding origin and each individual
 iso.prob contain relative probabilities for each breeding origin and individual
 iso.origin contains the likely/unlikely origins for each individual
}
\description{
Function to estimate the likelihood of origin and likely/unlikely origins for stable hydrogen isotope samples
}
\examples{
iso_assign(dd = woth_dd, df.base = woth_base$ddf)
}
\keyword{assignment;}
\keyword{isotopes}
\keyword{stable}

