% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wght_coord.R
\name{wght_coord}
\alias{wght_coord}
\title{Breeding assignment using weighted stable isotope and abundance data}
\usage{
wght_coord(summ, iso = TRUE)
}
\arguments{
\item{summ}{Dataframe containing the isotope- and abundance-base assignment results from the iso_assign() and/or abun_assign() function}

\item{iso}{Should coordinates be based on isotope assignment or abundance assignment?}
}
\value{
Data frame containing the estimate latitude & logitude and uncertainty for each individual.
}
\description{
Function to estimate the posterior probability of origin and likely/unlikely origins for stable hydrogen isotope samples and relative breeding abundance
}
\examples{
woth_coord <- wght_coord(summ = woth_iso)
}
\keyword{abundance}
\keyword{assignment;}
\keyword{isotopes;}
\keyword{stable}

