% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abun_assign.R
\name{abun_assign}
\alias{abun_assign}
\title{Breeding assignment using weighted stable isotope and abundance data}
\usage{
abun_assign(iso.like, rel.abun, iso.weight, abun.weight, odds = 3)
}
\arguments{
\item{rel.abun}{Vector of relative abundance values for each breeding cell (should sum to 1).}

\item{iso.weight}{Weight value to apply to stable isotope data.}

\item{abun.weight}{Weight value to apply to relative abundance data.}

\item{odds}{Denominator of the odds ratio for determining likely/unlikely origins. Default is 3.}

\item{iso.lik}{Dataframe or matrix containing isotope-based likelihood of origin for each breeding cell and individual.}
}
\value{
A list containing two dataframes: prob contain posterior probabilities for each breeding origin and individual & origin contains the likely/unlikely origins for each individual
}
\description{
Function to estimate the posterior probability of origin and likely/unlikely origins for stable hydrogen isotope samples and relative breeding abundance
}
\examples{
weight_assign(iso.lik = amre_iso$iso.prob, rel.abun = woth_base$rel.abun, iso.weight = -0.1, abun.weight = -0.9)
}
\keyword{abundance}
\keyword{assignment;}
\keyword{isotopes;}
\keyword{stable}

