% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/area_assign.R
\name{area_assign}
\alias{area_assign}
\title{Estimate the assignment area for known-origin tissue samples}
\usage{
area_assign(origin)
}
\arguments{
\item{origin}{Matrix with columns indicating likely/unlikely cells for each individual.}
}
\value{
Assignment area (i.e., the mean number of cells identified as likely for all individuals)
}
\description{
Function to estimate the mean number of cells identified as likely for all individuals in a given sample
}
\examples{
area_assign(origin = amre_assign$iso.origin, base = amre.base)
}
\keyword{area}
\keyword{assignment}
\keyword{assignment;}
\keyword{isotopes;}
\keyword{stable}

