% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_rate.R
\name{error_rate}
\alias{error_rate}
\title{Estimate the assignment error rate for known-origin tissue samples}
\usage{
error_rate(summ, origin_cell, iso = TRUE)
}
\arguments{
\item{summ}{Assignment results from either the iso_assign() or abun_assign() functions}

\item{origin_cell}{Vector containing the true breeding cell for each individual}

\item{iso}{Should area be estimated for the isotope-only assignment (default) or with abundance used a prior}
}
\value{
Error rate (i.e., the proportion of individuals that are miss classified)
}
\description{
Function to estimate the proportion of individuals that are incorrectly assigned to their actual breeding origin
}
\examples{
error_rate(summ = woth_iso, origin.cell = woth_origin_cell)
}
\keyword{assignment;}
\keyword{error}
\keyword{isotopes;}
\keyword{rate}
\keyword{stable}

